Preview server refactoring (#2497)
* Split preview server by preview and web server classes

* Support web server dynamic loading

* improve loggin for reload

* clean dead threads

* use webrick

* remove .tool-versions

* finish cleanup

* finish cleanup

* back original logger
